# shared/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Stellive Notification Hub API
  version: 0.1.0
  description: Unofficial fan-project API. Official YouTube live events for the Stellive official channel are excluded.
paths:
...
  /v1/hub-events:
    get:
      summary: Official-source goods and event feed with filters.
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [online_goods, online_collab, offline_concert, offline_collab, offline_popup, ticketing]
        - name: tag
          in: query
          schema:
            $ref: "#/components/schemas/HubEventTag"
        - name: participationMode
          in: query
          schema:
            type: string
            enum: [online, offline, hybrid]
        - name: status
          in: query
          schema:
            type: string
            enum: [announced, upcoming, open, closing_soon, ended, cancelled]
        - name: generationId
          in: query
          schema:
            type: string
        - name: memberId
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            oneOf:
              - type: string
              - type: integer
            default: 25
            maximum: 100
      responses:
        "200":
          description: HubEvent list response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HubEventListResponse"
        "400":
          description: Invalid HubEvent query.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HubEventQueryError"
  /v1/hub-events/calendar:
    get:
      summary: Hub event calendar grouped by local date.
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
        - name: timezone
          in: query
          schema:
            type: string
            default: Asia/Seoul
        - name: includeSpecialDays
          in: query
          schema:
            type: boolean
            default: true
        - name: entryKind
          in: query
          schema:
            type: string
            description: Comma-separated calendar entry kinds.
            example: hub_event,member_birthday
      responses:
        "200":
          description: Goods event calendar days.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HubCalendarResponse"
        "400":
          description: Invalid calendar query, date range, or timezone.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HubEventQueryError"
  /v1/hub-events/widget-snapshot:
    get:
      summary: Compact hub event calendar snapshot for mobile widgets.
      parameters:
        - name: timezone
          in: query
          schema:
            type: string
            default: Asia/Seoul
        - name: includeSpecialDays
          in: query
          schema:
            type: boolean
            default: true
        - name: entryKind
          in: query
          schema:
            type: string
            description: Comma-separated calendar entry kinds.
            example: hub_event,member_birthday
        - name: limit
          in: query
          schema:
            oneOf:
              - type: string
              - type: integer
            default: 5
            maximum: 10
      responses:
        "200":
          description: Compact widget snapshot.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HubCalendarWidgetSnapshot"
        "400":
          description: Invalid widget snapshot query.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HubEventQueryError"
  /v1/hub-events/{id}:
    get:
      summary: HubEvent detail by id.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: HubEvent detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HubEvent"
        "404":
          description: Hub event not found.
          content:
            application/json:
              schema:
                type: object
                required: [error]
                properties:
                  error:
                    type: string
                    enum: [hub_event_not_found]
...
components:
  securitySchemes:
    internalBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Bearer token protection for `/v1/internal/*` operational routes.
...
  schemas:
    HubEventListResponse:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/HubEvent"
        nextCursor:
          type: string
    HubEventQueryError:
      type: object
      required:
        - error
        - field
      properties:
        error:
          type: string
          enum:
            - invalid_hub_event_query
        field:
          type: string
    HubCalendarEntry:
      type: object
      required:
        - id
        - eventId
        - entryKind
        - title
        - category
        - tags
        - status
        - participationMode
        - generationId
        - displayDate
        - displayTimeText
        - sourceLabel
        - appDeepLink
      properties:
        id:
          type: string
        eventId:
          type: string
        entryKind:
          type: string
          enum:
            - hub_event
            - member_birthday
            - generation_anniversary
        specialDayKind:
          type: string
          enum:
            - member_birthday
            - generation_anniversary
        specialDayLabel:
          type: string
        scheduleItemId:
          type: string
        scheduleKind:
          $ref: "#/components/schemas/HubEventScheduleKind"
        scheduleLabel:
          type: string
        title:
          type: string
        displayTitle:
          type: string
        category:
          type: string
          enum:
            - online_goods
            - online_collab
            - offline_concert
            - offline_collab
            - offline_popup
            - ticketing
        tags:
          type: array
          items:
            $ref: "#/components/schemas/HubEventTag"
          example: [album]
        status:
          type: string
          enum:
            - announced
            - upcoming
            - open
            - closing_soon
            - ended
            - cancelled
        participationMode:
          type: string
          enum:
            - online
            - offline
            - hybrid
        generationId:
          type: string
        memberId:
          type: string
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        displayDate:
          type: string
          format: date
        displayTimeText:
          type: string
        sourceLabel:
          type: string
        appDeepLink:
          type: string
        platformUrl:
          type: string
    HubCalendarDay:
      type: object
      required:
        - date
        - entries
      properties:
        date:
          type: string
          format: date
        entries:
          type: array
          items:
            $ref: "#/components/schemas/HubCalendarEntry"
    HubCalendarResponse:
      type: object
      required:
        - timezone
        - from
        - to
        - days
      properties:
        timezone:
          type: string
        from:
          type: string
          format: date
        to:
          type: string
          format: date
        days:
          type: array
          items:
            $ref: "#/components/schemas/HubCalendarDay"
    HubCalendarWidgetSnapshot:
      type: object
      required:
        - generatedAt
        - timezone
        - entries
        - staleAfter
      properties:
        generatedAt:
          type: string
          format: date-time
        timezone:
          type: string
        entries:
          type: array
          items:
            $ref: "#/components/schemas/HubCalendarEntry"
      staleAfter:
        type: string
        format: date-time
...
    HubEventImagePolicyState:
      type: string
      enum:
        - none
        - official_runtime_url
        - third_party_allowed
        - verify_required
        - blocked
    HubEventImage:
      type: object
      required:
        - policyState
      properties:
        policyState:
          $ref: "#/components/schemas/HubEventImagePolicyState"
        url:
          type: string
          format: uri
        sourceLabel:
          type: string
        sourceUrl:
          type: string
          format: uri
        altText:
          type: string
    HubEventLinkKind:
      type: string
      enum:
        - source
        - purchase
        - ticket
        - reservation
        - content
        - video
        - map
        - custom
    HubEventLinkInput:
      type: object
      required:
        - kind
        - url
        - sortOrder
      properties:
        id:
          type: string
        kind:
          $ref: "#/components/schemas/HubEventLinkKind"
        label:
          type: string
          maxLength: 80
        url:
          type: string
          format: uri
          pattern: "^https://"
        sortOrder:
          type: integer
          minimum: 0
    HubEventLink:
      allOf:
        - $ref: "#/components/schemas/HubEventLinkInput"
        - type: object
          required:
            - id
            - createdAt
            - updatedAt
          properties:
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
    HubEvent:
      type: object
      required:
        - id
        - category
        - tags
        - participationMode
        - status
        - title
        - generationId
        - sourceUrl
        - sourceLabel
        - sourceType
        - notificationEligible
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            $ref: "#/components/schemas/HubEventTag"
          example: [album]
        participationMode:
          type: string
        status:
          type: string
        title:
          type: string
        summary:
          type: string
        generationId:
          type: string
        memberId:
          type: string
        sourceUrl:
          type: string
          format: uri
        sourceLabel:
          type: string
        sourceType:
          type: string
        scheduleMode:
          $ref: "#/components/schemas/HubEventScheduleMode"
        scheduleItems:
          type: array
          items:
            $ref: "#/components/schemas/HubEventScheduleItem"
        links:
          type: array
          items:
            $ref: "#/components/schemas/HubEventLink"
        announcedAt:
          type: string
          format: date-time
          nullable: true
        startsAt:
          type: string
          format: date-time
          nullable: true
        endsAt:
          type: string
          format: date-time
          nullable: true
        purchaseUrl:
          type: string
          format: uri
        ticketUrl:
          type: string
          format: uri
        venueName:
          type: string
        venueAddress:
          type: string
        image:
          $ref: "#/components/schemas/HubEventImage"
        notificationEligible:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    HubEventScheduleMode:
      type: string
      enum:
        - single_window
        - timeline
    HubEventScheduleKind:
      type: string
      enum:
        - main_window
        - announcement
        - sales_open
        - ticket_open
        - content_reveal
        - release
        - deadline
        - custom
    HubEventTimePrecision:
      type: string
      enum:
        - date
        - datetime
    HubEventScheduleItem:
      type: object
      required:
        - id
        - kind
        - label
        - startsAt
        - timePrecision
        - timezone
        - notificationEligible
        - isPrimary
        - sortOrder
      properties:
        id:
          type: string
        kind:
          $ref: "#/components/schemas/HubEventScheduleKind"
        title:
          type: string
          maxLength: 160
        label:
          type: string
          maxLength: 80
        description:
          type: string
          nullable: true
          maxLength: 2000
        startsAt:
          oneOf:
            - type: string
              format: date
            - type: string
              format: date-time
        endsAt:
          oneOf:
            - type: string
              format: date
            - type: string
              format: date-time
        timePrecision:
          $ref: "#/components/schemas/HubEventTimePrecision"
        timezone:
          type: string
          default: Asia/Seoul
        actionUrl:
          type: string
          format: uri
        sourceUrl:
          type: string
          format: uri
        sourceLabel:
          type: string
        links:
          type: array
          items:
            $ref: "#/components/schemas/HubEventLink"
        notificationEligible:
          type: boolean
        isPrimary:
          type: boolean
        sortOrder:
          type: integer
        cancelledAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    HubEventTag:
      type: string
      enum: [album]
...